infrastructure/ocr/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
pytesseract==0.3.10
pandas==2.1.4
openpyxl==3.1.2
PyPDF2==3.0.1
requests==2.31.0
